+++ nextapp/src/styles/layout.module.css
@@ -10,7 +10,3 @@
   right: 10px;
   top: 10px;
 }
-
-.dropdownstyle {
-  maxheight: 300;
-}
